feat: move games hub to /leikir, port Hörpuhopp to React, security hardening
- Consolidate all games under /leikir (was /skatathing); add 308 redirects
  for old paths so bookmarks keep working
- Port Hörpuhopp from static HTML/JS to a proper Next.js page: game engine
  extracted to gameEngine.ts, canvas managed in useEffect, leaderboard as
  React state (desktop sidebar / mobile bottom sheet); SSR disabled via
  dynamic import to avoid hydration mismatches
- Fix path traversal + XSS in [game]/route.ts via KNOWN_GAMES whitelist
  (lib/leikir-games.ts); same guard added to scores API proxy
- Add error handling for non-JSON backend responses in scores proxy
- Backend: cap scores at ge=1, le=999_999; add rate limiting (10/60s) on
  score submission; fix CORS on unhandled exceptions via global exception
  handler; rename heidursordla router prefix to /leikir/heidursordla
- Rename components/skatathing → components/leikir; split skatathing.module.css
  into leikir.module.css (hub) and heidursordla.module.css (game pages)
- HUD pills now size to their text content via measureText instead of fixed widths
- Leaderboard and page background use design tokens (theme-aware, dark mode)
