+++ frontend/components/leikir/heidursordla/HeidursordlaGameOver.tsx
