+++ frontend/components/leikir/heidursordla/HeidursordlaPlayView.tsx
