+++ frontend/components/leikir/horpuhopp/horpuhopp.module.css
@@ -1,0 +1,40 @@
+/* Page shell — fills the viewport below the sticky header.
+   Canvas + sidebar sit side-by-side on desktop; on mobile the canvas fills
+   the screen and the leaderboard slides up as a bottom sheet. */
+
+.root {
+  background: hsl(var(--sl-color-background));
+  /* Fill exactly the space below the header, no scrolling */
+  height: calc(100dvh - 60px);
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  overflow: hidden;
+}
+
+.gameArea {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  justify-content: center;
+  gap: var(--sl-spacing-lg);
+  padding: var(--sl-spacing-lg);
+  height: 100%;
+  width: 100%;
+  max-width: 760px;
+}
+
+@media (max-width: 639px) {
+  .root {
+    padding: 0;
+    align-items: flex-start;
+    /* Full dvh on mobile so canvas reaches the bottom edge */
+    height: calc(100dvh - 56px);
+  }
+
+  .gameArea {
+    padding: 0;
+    gap: 0;
+    max-width: 100%;
+  }
+}
