--- frontend/public/leikir/horpuhopp/style.css
@@ -1,24 +1,0 @@
-* {
-  margin: 0;
-  padding: 0;
-  box-sizing: border-box;
-}
-
-html,
-body {
-  width: 100%;
-  height: 100%;
-  background: hsl(160 42% 15%);
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  overflow: hidden;
-}
-
-#gameCanvas {
-  display: block;
-  border-radius: 10px;
-  box-shadow: 0 8px 40px rgba(0, 0, 0, 0.6);
-  max-width: 100vw;
-  max-height: 100vh;
-}
